requirements.txt
pytorch torchvision torchaudio pytorch-cuda -c pytorch -c nvidia
cudann
swig 
gymnasium[box2d]
matplotlib
numpy